+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Look up WC kubeconfig based on Cluster name instead of OpenStackCluster name.
+- Update changed DNS record values.
 
 ## [0.1.0] - 2021-12-15