+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2022-01-05
+
 ### Added
 
 - Add core Cluster to scope in addition to infrastructure cluster.
@@ -30,5 +32,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Create api and ingress entries in Route53.
 
-[Unreleased]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.1.1...HEAD
+[0.1.1]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/giantswarm/dns-operator-openstack/releases/tag/v0.1.0